move dcm_2d from test_idata to conftest
